- refactor navbar components and improve naming
