+++ src/components/Menu/CombinationsComponent.js
