+++ src/components/Menu/dishes/RiceComponent.js
