+++ src/components/Menu/dishes/VegetableAndAlmonds.js
