+++ src/components/Navbar/MobileMenuComponent.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Button, Header, Icon, Modal, Menu } from "semantic-ui-react";
-import MenuListComponenet from "../../menu/MenuListComponent";
+import MenuListComponenet from "../Menu/MenuListComponent";
 
 const MenuStyle = {
   top: "5%",
