+++ src/components/desktop/menu/DesktopMenuRoutesComponent.js
@@ -1,17 +1,17 @@
 import React, { useState } from "react";
 import { Sidebar, Segment, div, Item } from "semantic-ui-react";
 import { Route, Switch } from "react-router-dom";
-import CombinationsComponent from "../../menu/CombinationsComponent";
-import FullcourseComponenet from "../../menu/FullCourseComponent";
-import SoupComponent from "../../menu/dishes/SoupsComponent";
-import AppetizersComponent from "../../menu/dishes/AppetizersComponent";
-import RiceComponent from "../../menu/dishes/RiceComponent";
-import EggFooYoungComponent from "../../menu/dishes/EggFooYoungComponent";
-import ChowMeinComponent from "../../menu/dishes/ChowMeinComponent";
-import SweetAndSourComponent from "../../menu/dishes/Sweet&SourComponent";
-import VegetablesAndAlmondsComponent from "../../menu/dishes/VegetableAndAlmonds";
-import CantoneseComponent from "../../menu/dishes/CantoneseComponent";
-import CanadianComponent from "../../menu/CanadianDishesComponent";
+import CombinationsComponent from "../../Menu/CombinationsComponent";
+import FullcourseComponenet from "../../Menu/FullCourseComponent";
+import SoupComponent from "../../Menu/dishes/SoupsComponent";
+import AppetizersComponent from "../../Menu/dishes/AppetizersComponent";
+import RiceComponent from "../../Menu/dishes/RiceComponent";
+import EggFooYoungComponent from "../../Menu/dishes/EggFooYoungComponent";
+import ChowMeinComponent from "../../Menu/dishes/ChowMeinComponent";
+import SweetAndSourComponent from "../../Menu/dishes/Sweet&SourComponent";
+import VegetablesAndAlmondsComponent from "../../Menu/dishes/VegetableAndAlmonds";
+import CantoneseComponent from "../../Menu/dishes/CantoneseComponent";
+import CanadianComponent from "../../Menu/CanadianDishesComponent";
 const menuRouteStyle = {
   menuItemStyle: {
     paddingLeft: "50px",
